(PC-6701) split calendar in DayComponent
